refactor(7114): keep harness fixture responsibilities bounded
Move generic report and artifact I/O out of the agent-harness support module so canonical settlement construction remains isolated and every touched source file stays below the repository size limit.

Constraint: Preserve the now-green verifier compatibility behavior.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep canonical settlement evidence construction separate from generic fixture I/O.
Tested: 23 agent-harness contracts and targeted strict clippy.
Not-tested: Full workspace suite deferred to integration phase.
